Updates for M-Unit testing functionality
Changes to functionality include:

STARTUP and SHUTDOWN will be run in each routine instead of the first
instance of STARTUP and SHUTDOWN encountered being the only ones that
were run.  This change was made because of possible suites of tests
which have their own STARTUP and SHUTDOWN that need to be run.  Also, it
does not depend on the order in which test routines are run.

While the coverage functionality can only be run on GT.M systems at the
present time, the routine %utcover has been added to provide additional
ability to run a number of unit tests at one time so that coverage
across all of them can be obtained in one run, this includes
the ability to specify entry at different APIs (e.g., if both verbose
and nonverbose entries are available) as well as run code which is not
part of a unit test to include in the analysis.  The routine %uttcovr
uses this functionality to provide a means of doing coverage analysis on
the %ut routines and can be used as an example of its usage.

A new tag (COV) has been added to the %ut routine (it simply calls the
one in %ut1) so that all user APIs can be accessed within the %ut
routine.

A potential problem with setting up the %ut namespace in the local
instance was addressed by Sam Habiel in the preinstall routine ZZUTPRE.